mesh_builder: prwmbTexid takes in a uint32_t as a parameter

+++ include/mesh_builder.h
@@ -107,7 +107,7 @@ void prwmbIndex(PRWmeshBuilder* builder, size_t numIndicies, ...);
 
 void prwmbIndexv(PRWmeshBuilder* builder, size_t numIndicies, const uint32_t* indicies);
 
-void prwmbTexid(PRWmeshBuilder* builder, uint8_t texid);
+void prwmbTexid(PRWmeshBuilder* builder, uint32_t texid);
 
 prwvfVTXFMT* prwmbGetVertexFormat(PRWmeshBuilder* builder);
